project/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.26.1
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
psycopg2